backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
google-genai
pymupdf
openai
chromadb
requests
openpyxl
httpx

# Testing
pytest
pytest-cov
pytest-asyncio
